requirements.txt
anyio==3.7.0
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2
pydantic==1.10.9
PyJWT==2.7.0
PyMuPDF==1.22.3
pypdf==3.9.1
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
typing-extensions==4.6.3
uvicorn==0.22.0
